# [1.7.x] Do not fire `build/publishDiagnostics` if there are (and were) no problems